+++ src/Infrastructure/EventBus/Outbound/PublisherInterface.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * Copyright 2023 Cloud Creativity Limited
+ * Copyright 2024 Cloud Creativity Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,20 +17,17 @@
 
 declare(strict_types=1);
 
-namespace CloudCreativity\Modules\IntegrationEvents;
+namespace CloudCreativity\Modules\Infrastructure\EventBus\Outbound;
 
-use DateTimeImmutable;
-use Ramsey\Uuid\UuidInterface;
+use CloudCreativity\Modules\Infrastructure\EventBus\IntegrationEventInterface;
 
-interface IntegrationEventInterface
+interface PublisherInterface
 {
     /**
-     * @return UuidInterface
+     * Publish an outbound integration event.
+     *
+     * @param IntegrationEventInterface $event
+     * @return void
      */
-    public function uuid(): UuidInterface;
-
-    /**
-     * @return DateTimeImmutable
-     */
-    public function occurredAt(): DateTimeImmutable;
+    public function publish(IntegrationEventInterface $event): void;
 }