+++ tests/Unit/Infrastructure/EventBus/Middleware/LogOutboundIntegrationEventTest.php
@@ -19,8 +19,8 @@ declare(strict_types=1);
 
 namespace CloudCreativity\Modules\Tests\Unit\Infrastructure\EventBus\Middleware;
 
-use CloudCreativity\Modules\Infrastructure\EventBus\Middleware\LogIntegrationEventInbound;
-use CloudCreativity\Modules\IntegrationEvents\IntegrationEventInterface;
+use CloudCreativity\Modules\Infrastructure\EventBus\IntegrationEventInterface;
+use CloudCreativity\Modules\Infrastructure\EventBus\Middleware\LogOutboundIntegrationEvent;
 use CloudCreativity\Modules\Tests\Unit\Infrastructure\EventBus\TestIntegrationEvent;
 use CloudCreativity\Modules\Toolkit\ModuleBasename;
 use LogicException;
@@ -29,7 +29,7 @@ use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
 use Psr\Log\LogLevel;
 
-class LogIntegrationEventInboundTest extends TestCase
+class LogOutboundIntegrationEventTest extends TestCase
 {
     /**
      * @var LoggerInterface&MockObject
@@ -67,7 +67,7 @@ class LogIntegrationEventInboundTest extends TestCase
                 return true;
             });
 
-        $middleware = new LogIntegrationEventInbound($this->logger);
+        $middleware = new LogOutboundIntegrationEvent($this->logger);
         $middleware($this->event, function (IntegrationEventInterface $received): void {
             $this->assertSame($this->event, $received);
         });
@@ -78,8 +78,8 @@ class LogIntegrationEventInboundTest extends TestCase
         ];
 
         $this->assertSame([
-            [LogLevel::DEBUG, "Receiving integration event {$eventName}.", $context],
-            [LogLevel::INFO, "Received integration event {$eventName}.", $context],
+            [LogLevel::DEBUG, "Publishing integration event {$eventName}.", $context],
+            [LogLevel::INFO, "Published integration event {$eventName}.", $context],
         ], $logs);
     }
 
@@ -99,7 +99,7 @@ class LogIntegrationEventInboundTest extends TestCase
                 return true;
             });
 
-        $middleware = new LogIntegrationEventInbound($this->logger, LogLevel::NOTICE, LogLevel::WARNING);
+        $middleware = new LogOutboundIntegrationEvent($this->logger, LogLevel::NOTICE, LogLevel::WARNING);
         $middleware($this->event, function (IntegrationEventInterface $received) {
             $this->assertSame($this->event, $received);
         });
@@ -110,8 +110,8 @@ class LogIntegrationEventInboundTest extends TestCase
         ];
 
         $this->assertSame([
-            [LogLevel::NOTICE, "Receiving integration event {$eventName}.", $context],
-            [LogLevel::WARNING, "Received integration event {$eventName}.", $context],
+            [LogLevel::NOTICE, "Publishing integration event {$eventName}.", $context],
+            [LogLevel::WARNING, "Published integration event {$eventName}.", $context],
         ], $logs);
     }
 
@@ -131,9 +131,9 @@ class LogIntegrationEventInboundTest extends TestCase
         $this->logger
             ->expects($this->once())
             ->method('log')
-            ->with(LogLevel::DEBUG, "Receiving integration event {$eventName}.", $context);
+            ->with(LogLevel::DEBUG, "Publishing integration event {$eventName}.", $context);
 
-        $middleware = new LogIntegrationEventInbound($this->logger);
+        $middleware = new LogOutboundIntegrationEvent($this->logger);
 
         try {
             $middleware($this->event, static function () use ($expected) {